+++ CHANGELOG.md
@@ -3,13 +3,14 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.35.3] - 2017-06-19
 ---------------------
 ##### Changed
 - Reduced swipe sensitivity to avoid accidental "click-swiping" with a mouse.
 
 ##### Fixed
 - Page swipe bypasses block-page-navigation-on-new-constraint-error feature.
+- If form in Pages mode has only one page, this page is not shown.
 
 [4.35.2] - 2017-06-13
 ---------------------